restructure of file locations
