+++ src/components/RowWithDelete/RowWithDelete.js
