+++ src/components/TypeComponents/SimpleSearch/Form/configurableProperties/SimpleSearchConfigurableProperties.js
