+++ src/components/TypeComponents/SimpleSearch/Form/filters/SimpleSearchFilterArray.js
