+++ src/components/TypeComponents/SimpleSearch/Form/filters/SimpleSearchFilterRuleArray.js
