+++ src/components/TypeComponents/SimpleSearch/Form/formParsing/submitWithTokens.js
