+++ src/components/TypeComponents/SimpleSearch/Form/results/SimpleSearchResultField.js
