+++ src/components/TypeComponents/SimpleSearch/Form/results/SimpleSearchResults.css
