+++ src/components/TypeComponents/SimpleSearch/Widget/SimpleSearch.css
