+++ src/components/TypeComponents/SimpleSearch/Widget/index.js
