+++ src/components/TypeComponents/SimpleSearch/Widget/simpleSearchColumnParser.js
