+++ src/components/Widget/Widget.js
