+++ src/components/Widget/WidgetFooter.js
