+++ src/components/Widget/WidgetFooter.test.js
