+++ src/components/Widget/index.js
